+++ EwkQxObd.Core/Model/Views/Syngio.cs
@@ -12,7 +12,7 @@ namespace EwkQxObd.Core.Model.Views
     public partial class Syngio
     {
         [Column("InstrumentId", TypeName = "bigint")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Column("System", TypeName = "char(16)")]
         public string System { get; set; } = string.Empty;
@@ -24,7 +24,7 @@ namespace EwkQxObd.Core.Model.Views
         public string? NetworkName { get; set; } = string.Empty;
 
         [Column("NetworkId", TypeName = "bigint")]
-        public long? NetworkId { get; set; }
+        public int? NetworkId { get; set; }
 
         [Column("InstrumentGroup", TypeName = "nvarchar(64)")]
         public string? InstrumentGroup { get; set; } = string.Empty;
